💚 *.yml: Fixed CI build
For some reasons, `npm@6.5.x` seems to break on UNIX system when running
the command `npm cit` but it works fine on my testing machine with
Windows 10 x64. This is totally strange and has never happened to me at
least. Now reverting to good old commands of installing deps and
running tests