🔥 packge-lock.json: Forced clean local NPM cache to rebuild lockfile
There is a unpublished package `har-validator@5.1.2` but NPM still
installing that missing version and not updating the lockfile which
caused such error for, say CI, to keep grabbing something that is no
longer in the NPM registry. But it still in the local cache of my
machine thus resulting in such chaos.

FYI, see anmonteiro/lumo/pull/456 to learn about the said issue.